requirements.txt
asgiref==3.8.1
Django==5.1
django-allauth==0.54.0
django-crispy-forms==1.14.0
oauthlib==3.2.2
pillow==10.4.0
PyJWT==2.9.0
python3-openid==3.2.0
requests-oauthlib==2.0.0
sqlparse==0.5.1
